refactor(config): rename `config/enums.py` to `config/namespace.py`
