+++ src/declusor/config/__init__.py
@@ -1,4 +1,4 @@
-from .enums import FileFunc, Language
+from .namespace import FileFunc, Language, DeclusorOptions
 from .exceptions import ControllerError, DeclusorException, ExitRequest, InvalidOperation, ParserError, PromptError, RouterError
 from .settings import BasePath, Settings
 
@@ -6,6 +6,7 @@ __all__ = [
     "BasePath",
     "ControllerError",
     "DeclusorException",
+    "DeclusorOptions",
     "ExitRequest",
     "FileFunc",
     "InvalidOperation",
