+++ src/declusor/config/namespace.py
